+++ test/scheduler/test_InferenceManager.cpp
@@ -40,7 +40,7 @@ TEST_P(InferenceManagerTest, Simple){
 
     PrePostProcessor pp_processor(test_params.inference_config);
     BackendBase* custom_processor = nullptr; // Use default processor
-    ContextConfig context_config;
+    ContextConfig context_config(2); // Use 2 threads for testing, gh runner max on macOS is 3
 
     InferenceManager inference_manager(
         pp_processor,
@@ -65,7 +65,7 @@ TEST_P(InferenceManagerTest, WithCustomLatency) {
 
     PrePostProcessor pp_processor(test_params.inference_config);
     BackendBase* custom_processor = nullptr; // Use default processor
-    ContextConfig context_config;
+    ContextConfig context_config(2); // Use 2 threads for testing, gh runner max on macOS is 3
 
     InferenceManager inference_manager(
         pp_processor,
@@ -102,7 +102,7 @@ TEST_P(InferenceManagerTest, WithEmptyCustomLatency) {
 
     PrePostProcessor pp_processor(test_params.inference_config);
     BackendBase* custom_processor = nullptr; // Use default processor
-    ContextConfig context_config;
+    ContextConfig context_config(2); // Use 2 threads for testing, gh runner max on macOS is 3
 
     InferenceManager inference_manager(
         pp_processor,
@@ -140,7 +140,7 @@ TEST_P(InferenceManagerTest, WithPartialCustomLatency) {
 
     PrePostProcessor pp_processor(test_params.inference_config);
     BackendBase* custom_processor = nullptr; // Use default processor
-    ContextConfig context_config;
+    ContextConfig context_config(2); // Use 2 threads for testing, gh runner max on macOS is 3
 
     InferenceManager inference_manager(
         pp_processor,
@@ -217,25 +217,119 @@ INSTANTIATE_TEST_SUITE_P(
             InferenceConfig(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 2048}}, {{1, 1, 2048}})},
-                1.f
+                1.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 4095 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000, true),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1024}}, {{360}})},
+                ProcessingSpec({1}, {2}, {160}, {1}),
+                3.f,
+                0,
+                false,
+                0.5f,
+                2
+            ),
+            { 1 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000, true),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1024}}, {{360}})},
+                ProcessingSpec({1}, {2}, {160}, {1}),
+                5.f,
+                0,
+                false,
+                0.5f,
+                2
+            ),
+            { 5 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000, true),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1024}}, {{360}})},
+                ProcessingSpec({1}, {2}, {160}, {1}),
+                4.9f,
+                0,
+                false,
+                0.5f,
+                1
+            ),
+            { 20 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000, true),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1024}}, {{360}})},
+                ProcessingSpec({1}, {2}, {160}, {1}),
+                4.f,
+                0,
+                false,
+                0.f,
+                2
+            ),
+            { 12 }
         },
         InferenceManagerTestParams {
             HostConfig(1, 48000.0/2048, true),
             InferenceConfig(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 1}}, {{1, 1, 2048}})},
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
-            { 5886 }
+            { 5885 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000.0, true),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1, 2048}}, {{1, 1, 1}})},
+                40.f,
+                0,
+                false,
+                0.f,
+                2
+            ),
+            { 1 }
+        },
+        InferenceManagerTestParams {
+            HostConfig(2048, 48000.0, false),
+            InferenceConfig(
+                std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
+                std::vector<TensorShape>{TensorShape({{1, 1, 2048}}, {{1, 1, 1}})},
+                39.f,
+                0,
+                false,
+                0.f,
+                2
+            ),
+            { 1 }
         },
         InferenceManagerTestParams {
             HostConfig(1, 48000.0/2048),
             InferenceConfig(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 1}}, {{1, 1, 2048}})},
-                50.f
+                50.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 4096 },
         },
@@ -245,9 +339,13 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 2048}}, {{1, 4, 1}})},
                 ProcessingSpec({1}, {4}),
-                20.f
+                20.f,
+                0,
+                false,
+                0.f,
+                2
             ),
-            { 2 }
+            { 1 }
         },
         InferenceManagerTestParams {
             HostConfig(1./256., 48000./2048.),
@@ -255,7 +353,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 4, 1}}, {{1, 1, 2048}})},
                 ProcessingSpec({4}, {1}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 3960 }
         },
@@ -265,7 +367,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 16, 1}}, {{1, 1, 2048}})},
                 ProcessingSpec({16}, {1}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 2048 }
         },
@@ -275,7 +381,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 16, 1}}, {{1, 1, 2048}, {2, 256}})},
                 ProcessingSpec({16}, {1, 2}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 6144, 768 }
         },
@@ -285,7 +395,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 16, 1}, {2, 256}}, {{1, 1, 2048}, {3, 128}})},
                 ProcessingSpec({16, 2}, {1, 3}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 2048, 128 }
         },
@@ -295,7 +409,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 16, 1}, {2, 256}}, {{1, 1, 2048}, {3, 128}})},
                 ProcessingSpec({16, 2}, {1, 3}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 6144, 384 }
         },
@@ -305,7 +423,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 16, 1}, {2, 256}}, {{1, 2, 2048}})},
                 ProcessingSpec({16, 2}, {2}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 4924 }
         },
@@ -315,7 +437,11 @@ INSTANTIATE_TEST_SUITE_P(
             InferenceConfig(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 2048}}, {{1, 1, 2048}})},
-                13.f
+                13.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 2744 }
         },
@@ -324,7 +450,11 @@ INSTANTIATE_TEST_SUITE_P(
             InferenceConfig(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 1, 1024}}, {{1, 1, 1024}})},
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 2820 }
         },
@@ -334,7 +464,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 8, 1}}, {{1, 1, 1024}})},
                 ProcessingSpec({8}, {1}),
-                12.f
+                12.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 3072 }
         },
@@ -344,7 +478,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 8, 1}}, {{1, 1, 1024}})},
                 ProcessingSpec({8}, {1}),
-                13.f
+                13.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 3583 }
         },
@@ -354,9 +492,13 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 4, 1}, {2, 128}}, {{1, 1, 512}, {3, 64}})},
                 ProcessingSpec({4, 2}, {1, 1}),
-                50.f
+                50.f,
+                0,
+                false,
+                0.f,
+                2
             ),
-            { 18944, 7104 }
+            { 12800, 4800 }
         },
         InferenceManagerTestParams {
             HostConfig(1500, 44100./8., true, 1),
@@ -364,9 +506,13 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 4, 1}, {2, 128}, {1, 2}}, {{1, 1, 512}, {3, 64}})},
                 ProcessingSpec({4, 2, 1}, {1, 1}),
-                50.f
+                50.f,
+                0,
+                false,
+                0.f,
+                2
             ),
-            { 18944, 7104 }
+            { 12800, 4800 }
         },
         InferenceManagerTestParams {
             HostConfig(256., 48000./8, false, 1),
@@ -374,7 +520,11 @@ INSTANTIATE_TEST_SUITE_P(
                 std::vector<ModelData>{ModelData("placeholder", anira::InferenceBackend::CUSTOM)},
                 std::vector<TensorShape>{TensorShape({{1, 4, 1}, {2, 256}}, {{1, 1, 2048}, {3, 128}})},
                 ProcessingSpec({1, 2}, {1, 1}, {0, 256}, {2048, 0}),
-                40.f
+                40.f,
+                0,
+                false,
+                0.f,
+                2
             ),
             { 2048, 0 }
         }